aflplusplus: init 2.62c
Initial implementation for the aflplusplus port of afl.
Currently only the 'source fuzzing' mode is built,
but it shouldn't be too hard to enable binary fuzzing
using qemu as well.

To enable binary fuzzing for QEMU, a patch for
the build_qemu_support.sh script is needed that disables
checking the /usr/lib paths and instead uses another approach,
e.g. by using `pkg-config --validate glib-2.0`.

But that's for another time, and I'm personally not using
the QEMU mode of AFL.