buildPythonPackage: Fix after #112276
Fix #114952

Before it was using the wrong `makeWrapper`, the one that expected to
run on the target platform and produce packages for the "super-target".

The "super target" (sliding off the end of the platforms list) is
defined to be the same as the target, to cap it with a loop so there is
no sliding off. However, `targetPackages.targetPackages` is *not*
similarly defined. Per the bottom of `pkgs/stdenv/booter.nix` it is an
almost-empty package set that just exists as a hack to help some things
with exotic `depsTargetTarget`.

In particular, that rump "super `targetPackages`" does not contain a
`runtimeShell`, and that's the source of our eval-time error.

(cherry picked from commit fc74ad444af06291f367a6d50dd70f010392cd33)