libwacom: support for cross-compilation
Problem is that the generate-hwdb tool was not built for the build
platform, causing errors that the binary couldn't be executed.

This patch explains to meson that this tool must be built for the build
platform.

See upstream issue:
https://github.com/linuxwacom/libwacom/issues/365

This works for now, pull in whatever upstream decides to do later.