librsvg: support cross-compilation (broken-ish)
Note that this may break some things, as the loaders cache is not built
anymore now.

It would be better if we could convince gdk-pixbuf to build the cache
file, because we can then also generate it here and merge the two.
It doesn't seem very simple to do that though.