llvm_12: fix cross
Previously, the linker would be asked to link llvm-config (which is
built for the build system*) against the host system's ncurses, by
absolute path, resulting in a failed build.  I've added a patch to fix
LLVM and have it link against the appropriate curses library by name
instead, so that it'll use the normal Nixpkgs logic for library
discovery.

But unfortunately, that's not the whole story, because I couldn't get
the linker to actually look in the right place using normal Nixpkgs
ways of specifying dependencies.  The build cc wrapper doesn't add the
-L flag for the build ncurses if it's in nativeBuildInputs of
pkgsBuildBuild.  So in addition to my LLVM patch, I've also had to use
the rather obscure platform-specific NIX_LDFLAGS_* override variable,
to get the build system linker to look in the right place for ncurses.

* I'm using Nixpkgs/GNU terminology here, not LLVM terminology.

(cherry picked from commit 90995a9bb3beffd57b9214d06a64498030851f6a)