pkg-config: always prefix with target platform
This reduces the differences between native and cross-compilation.
Due to that it should help preventing some regressions related
to the naming of pkg-config when doing native vs cross-compilation.

Note that some packages may expect pkg-config to not be prefixed,
those will need to be fixed to refer to $PKG_CONFIG rather than
to the hardcoded name pkg-config.

Due to pkg-config not being in the bootstrap packages,
it's relatively easy to change.
For other packages (binutils, gcc) where this change should also
happen, it's a bit harder since the bootstrap binaries are (not yet)
prefixed.